+++ breeze-core/src/main/kotlin/core/extension/RegexExtensions.kt
@@ -46,7 +46,7 @@ fun Regex.Companion.fromRange(first: Int, last: Int): String {
 			}
 			//2-9, 10-13 -> [2-9]|1[0-3]
 			pairs.joinToString("|") { (a, b) ->
-				a.toString().toCharArray().zip(b.toString().toCharArray()) { a1, b1 ->
+				a.toString().zip(b.toString()) { a1, b1 ->
 					if(a1 == b1) "$a1" else "[$a1-$b1]"
 				}.joinToString("")
 			}